feat: decoupled transaction exclusion api from coordinator package
