--- coordinator/persistence/rejectedtransaction/src/main/kotlin/net/consensys/zkevm/persistence/dao/rejectedtransaction/RejectedTransactionsDao.kt
@@ -1,17 +1,0 @@
-package net.consensys.zkevm.persistence.dao.rejectedtransaction
-
-import kotlinx.datetime.Instant
-import net.consensys.linea.RejectedTransaction
-import tech.pegasys.teku.infrastructure.async.SafeFuture
-
-interface RejectedTransactionsDao {
-  fun saveNewRejectedTransaction(rejectedTransaction: RejectedTransaction): SafeFuture<Unit>
-
-  fun findRejectedTransactionByTxHash(
-    txHash: ByteArray
-  ): SafeFuture<RejectedTransaction?>
-
-  fun deleteRejectedTransactionsBeforeTimestamp(
-    timestamp: Instant
-  ): SafeFuture<Int>
-}
