+++ jvm-libs/linea/core/domain-models/build.gradle
@@ -5,10 +5,6 @@ plugins {
 dependencies {
   implementation project(":jvm-libs:generic:extensions:kotlin")
   implementation "org.apache.tuweni:tuweni-units:${libs.versions.tuweni.get()}"
-  implementation "com.fasterxml.jackson.core:jackson-annotations:${libs.versions.jackson.get()}"
-  implementation "com.fasterxml.jackson.core:jackson-databind:${libs.versions.jackson.get()}"
-  implementation "com.fasterxml.jackson.module:jackson-module-kotlin:${libs.versions.jackson.get()}"
-  implementation "com.fasterxml.jackson.datatype:jackson-datatype-jsr310:${libs.versions.jackson.get()}"
 }
 
 jar {
