+++ transaction-exclusion-api/app/src/main/kotlin/net/consensys/linea/transactionexclusion/repository/RejectedTransactionsRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package net.consensys.linea.transactionexclusion.repository
 
 import kotlinx.datetime.Instant
-import net.consensys.linea.RejectedTransaction
+import net.consensys.linea.transactionexclusion.RejectedTransaction
 import net.consensys.linea.transactionexclusion.RejectedTransactionsRepository
 import net.consensys.zkevm.persistence.dao.rejectedtransaction.RejectedTransactionsDao
 import tech.pegasys.teku.infrastructure.async.SafeFuture
