+++ transaction-exclusion-api/core/src/main/kotlin/net/consensys/linea/transactionexclusion/RejectedTransaction.kt
