+++ transaction-exclusion-api/core/src/main/kotlin/net/consensys/linea/transactionexclusion/RejectedTransactionsRepository.kt
@@ -1,7 +1,6 @@
 package net.consensys.linea.transactionexclusion
 
 import kotlinx.datetime.Instant
-import net.consensys.linea.RejectedTransaction
 import tech.pegasys.teku.infrastructure.async.SafeFuture
 
 interface RejectedTransactionsRepository {
