+++ transaction-exclusion-api/persistence/rejectedtransaction/src/integrationTest/resources/log4j2.xml
